requirements.txt
langgraph>=0.0.39
langchain>=0.2.1
langchain-core>=0.3.0
langchainhub>=0.1.15
langchain-openai>=0.1.7
langchain-community>=0.2.1
langserve>=0.0.51
openai>=1.30.1
tavily-python>=0.3.2
python-dotenv>=1.0.1